+++ modules/13_sweep_optimization/include/f700f/sweep_runner.hpp
@@ -82,5 +82,7 @@ std::shared_ptr<IChannelFactory> make_frequency_offset_channel_factory();
 std::shared_ptr<IChannelFactory> make_simple_gain_fading_channel_factory();
 
 SweepConfig make_m1_baseline_smoke_sweep_config(std::string output_directory);
+SweepConfig make_m2_channel_matrix_smoke_sweep_config(std::string output_directory);
+SweepConfig make_m2_channel_matrix_full_sweep_config(std::string output_directory);
 
 } // namespace f700f